% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvesearch.R
\name{getCVEDBInfo}
\alias{getCVEDBInfo}
\title{Gets information about CVE database}
\usage{
getCVEDBInfo(endpoint = "http://cve.circl.lu/api/dbInfo")
}
\arguments{
\item{endpoint}{URL}
}
\value{
information about CVE database
}
\description{
Gets information about CVE database
}
\examples{
getCVEDBInfo()
}

