% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvesearch.R
\name{getCVEProducts}
\alias{getCVEProducts}
\title{Gets a list of vendor products}
\usage{
getCVEProducts(vendor, endpoint = "http://cve.circl.lu/api/browse")
}
\arguments{
\item{vendor}{character string}

\item{endpoint}{URL}
}
\value{
A list of products for \code{vendor}
}
\description{
Gets a list of vendor products
}
\examples{
getVendorProducts('microsoft')
}

