% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvesearch.R
\name{getCVEs}
\alias{getCVEs}
\title{Gets a list of CVEs for given vendor/product combo}
\usage{
getCVEs(vendor, product, endpoint = "http://cve.circl.lu/api/search")
}
\arguments{
\item{vendor}{character string}

\item{product}{character string}

\item{endpoint}{URL}
}
\value{
A list of CVEs for given vendor/product combo
}
\description{
Gets a list of CVEs for given vendor/product combo
}
\examples{
getCVEs("microsoft", "office")
}

