% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvesearch.R
\name{getCVEVendors}
\alias{getCVEVendors}
\title{Gets a list of CVE vendors}
\usage{
getCVEVendors(endpoint = "http://cve.circl.lu/api/browse")
}
\arguments{
\item{endpoint}{URL}
}
\value{
A list of CVE vendors
}
\description{
Gets a list of CVE vendors
}
\examples{
getCVEVendors()
}

