% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvesearch.R
\name{getLatestCVEs}
\alias{getLatestCVEs}
\title{Gets a list of latest CVEs}
\usage{
getLatestCVEs(endpoint = "http://cve.circl.lu/api/last")
}
\arguments{
\item{endpoint}{URL}
}
\value{
A list of latest CVEs
}
\description{
Gets a list of latest CVEs
}
\examples{
getLatestCVEs()
}

